+++ lib/plugins/login/lib/getCliLoginById.js
@@ -1,0 +1,18 @@
+const gql = require('graphql-tag');
+
+module.exports = (id, apolloQueryFn) =>
+  apolloQueryFn({
+    fetchPolicy: 'network-only',
+    query: gql`
+      query cliLoginById($id: String!) {
+        cliLoginById(id: $id) {
+          encryptedAccessToken
+          encryptedIdToken
+          encryptedRefreshToken
+          encryptedKey
+          encryptedIv
+        }
+      }
+    `,
+    variables: { id },
+  }).then(response => response.data);
