+++ lib/plugins/login/login.js
@@ -3,7 +3,6 @@
 const BbPromise = require('bluebird');
 const jwtDecode = require('jwt-decode');
 const chalk = require('chalk');
-const gql = require('graphql-tag');
 const uuid = require('uuid');
 const has = require('lodash/has');
 const forge = require('node-forge');
@@ -15,6 +14,7 @@ const userStats = require('../../utils/userStats');
 const setConfig = require('../../utils/config').set;
 const createApolloClient = require('../../utils/createApolloClient');
 const decryptToken = require('./lib/decryptToken');
+const getCliLoginById = require('./lib/getCliLoginById');
 
 const config = {
   PLATFORM_FRONTEND_BASE_URL: 'https://platform.serverless.com/',
@@ -22,25 +22,6 @@ const config = {
 };
 
 const client = createApolloClient(config.GRAPHQL_ENDPOINT_URL);
-
-const getCliLoginById = id =>
-  client
-    .query({
-      fetchPolicy: 'network-only',
-      query: gql`
-        query cliLoginById($id: String!) {
-          cliLoginById(id: $id) {
-            encryptedAccessToken
-            encryptedIdToken
-            encryptedRefreshToken
-            encryptedKey
-            encryptedIv
-          }
-        }
-      `,
-      variables: { id },
-    })
-    .then(response => response.data);
 
 class Login {
   constructor(serverless, options) {
@@ -96,7 +77,7 @@ class Login {
           this.serverless.cli.log('Waiting for a successful authentication in the browser.');
           fetchTries = 0;
         }
-        getCliLoginById(cliLoginId)
+        getCliLoginById(cliLoginId, client.query)
           .then(response => {
             const hasAllToken =
               has(response, ['cliLoginById', 'encryptedAccessToken']) &&
